Partial/Pipe/Compose
Basic implementations of partial/pipe/compose function utilities.

Partial is useful for applying less than the total number of expected
arguments returning a new function that accepts the rest. This is a core
util for composition without auto currying everything.

Pipe/Compose are standard function composition tools to thread values
through many functions. Both expect all functions in the sequence to be
unary where some allow the first applied to be multi-arity and all
subsequent to be unary.